+++ client/include/autd3.hpp
@@ -3,7 +3,7 @@
 // Created Date: 14/04/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 21/11/2021
+// Last Modified: 22/11/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -21,7 +21,6 @@
 #include "autd3/sequence/primitive.hpp"
 
 namespace autd {
-using core::DataArray;
 using core::Geometry;
 using core::LinkPtr;
 using core::Matrix4X4;