+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Change transducers' geometry accessor
     * Introduce `Transducer` struct which contains id and position of transducer 
 * Introduce `Drive` struct which contains explicit phase and duty parameter for transducer
+* Change delay and duty offset api
+    * Delete `AUTDSetDutyOffset` and `AUTDSetOutputDelay` from capi
 * Delete `Geometry::device_idx_for_trans_idx`, use `Geometry::global_to_local_idx` instead
     * Delete `AUTDDeviceIdxForTransIdx` from capi
 * Delete `AUTDNumTransducers`, `AUTDTransPositionByGlobal`, `AUTDSendGain`, and `AUTDSendModulation` from capi